+++ DICUI/DICUI.csproj
@@ -99,6 +99,7 @@
     <Compile Include="UI\KnownSystemComboBoxItem.cs" />
     <Compile Include="UI\AllowedSpeeds.cs" />
     <Compile Include="Utilities\DumpEnvironment.cs" />
+    <Compile Include="Utilities\Extensions.cs" />
     <Compile Include="Utilities\Validators.cs" />
     <Compile Include="Utilities\Converters.cs" />
     <Compile Include="OptionsWindow.xaml.cs">
